requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0  # 本地测试需要，Railway不需要但加上无妨